# apply fused moe gate in ds v3/r1