@lambert0312 2025-04-14
> Not dependent; the main branch does not enable shared experts fusion by default, so this PR can be merged directly.

Ok